requirements.txt
Flask>=2.2
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
SQLAlchemy>=2.0
pyodbc>=5.0
Flask-Login>=0.6
python-dotenv>=1.0
waitress>=2.1
